[REF] knowledge: refactor permission panel
Following recent changes done to the performance of Knowledge's permission panel
members' computations. We are refactoring the permission panel to simplify it
and changing a bit its behavior.

Here's what's been done:
- no more check that "based on" articles are accessible by the user (more
coherent with the topbar "hierarchy").
- when the owner of a private article leaves the article, the article is
sent to the trash instead of archiving it (as suggested by the confirmation
dialog).
- portal users with permission "none" are not hidden anymore.
- the permission selectors are now dropdowns (more flexible than selectors
and has a style that is consistent with the app).
- the "remove member" button has been replaced by an option in the permission
dropdown.
- use record object as much as possible for field values instead of making RPC
calls to the DB

The permission panel is now a popover instead of a bootstrap drodpown.

Task-4273102

Part-of: odoo/enterprise#76261
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>