[FIX] hr_recruitment: company_id candidate applicant mismatch
Since applicants are aggregated into a candidate and the field
company_id is not related between these two models. If the company is
modified on a records, it won't match the one of the other model and so
will create access right issues.

To avoid this to hapend, we synchronize the company id between these tow
models when we modify one and we use/create a candidate with the same
company of the applicant on creation

We also propagate the company from the applicant to the candidate, when
the candidate is manually created from the backend.

task-4350838

closes odoo/enterprise#75091

X-original-commit: 571b7dd0a4874407501bcc39694d11158edb8bdd
Related: odoo/odoo#189503
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>