[REF] *: sanitize run function in tours
In commit, we remove empty runs that no longer have a reason to exist.
We also prefer to use tour helpers ( hover ) instead of using
makeVisible functions in order to simulate user interactions as best as
possible.
We take advantage of this commit to remove the nextTick helper in
webStudio (we no longer need it since 3bfd2e9cf8 )

closes odoo/enterprise#74641

X-original-commit: 2c84bb294a0cd8c10590b42cdfcc3c4be83607be
Related: odoo/odoo#188723
Signed-off-by: Bastien Pierre (ipb) <ipb@odoo.com>
Signed-off-by: Pierre Pulinckx (pipu) <pipu@odoo.com>