[IMP] documents: maximize the document preview to full screen
We maximize the document preview to full screen and add all the actions in the
header allowing the user to perform actions directly from the preview.

When the chatter is visible, we reduce the width so that the chatter remains
visible because this is where information about the document will be displayed.

Task-4306398

closes odoo/enterprise#74799

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>