[FIX] account_reports: properly handle groupby in account_codes engine for grouping keys used with multiple accounts
// TO REPRODUCE

1) Create a report with a single line, whose computation uses the account_codes engine, with formula "1". Set "partner_id" as the groupby value for that line. Setup a "load more limit" of 2 on your report.

2) Create 3 accounts, with codes 11, 12, 13

3) Create 3 partners: A, B, and C

4) Create journal entries, with the following lines (all balancing lines must use other accounts as the ones defined in 2) ), in that order:
    - on account 11, with partner A, debit=10
    - on account 11, with partner A, debit=20
    - on account 12, with partner A, debit=25
    - on account 11, without partner, debit=30
    - on account 11, without partner, debit=40
    - on account 12, with partner B, debit=50
    - on account 13, with partner C, debit=60

5) Open the report for a period covering all the created journal entries, and have a look at the sublines generated for the groupby.

    => With the limit being 2, you'd expect to see "A", "B", and a "Load more..." line allowing to visualize the rest (namely a line for "C", and "Unknown"). Instead of that, you see no "Load more ...", and line "A"'s amount is wrong.

// EXPLANATION

The SQL query of the account_codes engine is grouping not only by groupby key (here, partner_id), but also by account_id, to be able to process the "C" and "D" suffixes of the formula later on. Because of that, the LIMIT passed by the load_more_limit is applied on that grouping.

With a load_more_limit of 2, we try to load 3 lines. If we manage to load 3, we only display 2, but know we need to display the "Load more..." line. Else, we know there are no more results, so no need for that additional line.

So, with our example, the query runs with a LIMIT of 3, returning
    - 30 for A on account 11
    - 25 for A on account 12
    - 70 for no partner on account 11

Then, another issue happens:

The results returned by the SQL query (hence, by the engine) contain two times the same partner_id. From an engine perspective, the grouping key is only partner_id, not account_id (which is only used for an inner computation, with the SQL query). So that means the same grouping key has two different values. This is not legit, and causes one of them to shadow the other.

The report computation then receives 1 result for A, and 1 for "Unknown", so it renders them, and since there is no third result, it does not display the "Load more ...".

closes odoo/enterprise#75234

X-original-commit: e26cefc56cd922516849b97d9cd58bd10a2b927c
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>