[ADD] l10n_br_edi_stock: add transport information to the NF-e
NF-e invoices can optionally include sent packages. This integrates
with the stock module to provide a way for users to specify packages
on invoices.

Typically EDI is done before products are sent out ("ordered quantity"
invoicing policy). To handle this the user can enable 2-step
picking. In the first step they can record how they will package the
products. After that, the packages can be included on the invoice for
EDI. At a later time products can be delivered to the customer in the
last step. We provide a picking smart button in the invoice form view
to easily navigate to the related pickings to sort out the
packaging. This is typically done by the Faturista, a dedicated person
who handles everything related to EDI.

Note that for products to be included in packages they must be
storable (is_storable). The stock module allows to package
non-storable products as well, but in those cases the resulting
stock.quant.package's are empty (no quant_ids).

task-4298921

closes odoo/enterprise#78475

X-original-commit: 54af8e63be971b6565296a44c33ca65a575f2a62
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>