[FIX] timesheet_grid: hide Statusbar for the Portal User
In the community PR, we have used the timesheet_view_form_portal_user form
view for the portal user. To ensure that the status bar is not visible to
the portal user, we added a group restriction.

An alternative approach would be to hide the status bar based on context,
but I did not find it suitable. Therefore, we opted to use the group restriction.

task-4369891

closes odoo/enterprise#85270

X-original-commit: de4b1a017ee4e8316cb7a24b4fe9f3a2e083716b
Related: odoo/odoo#209457
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>