[FIX] documents: maintain owner access after pinning
Follow-up of 4d653bf5 where we forgot to make sure that the
owner keeps an `edit` permission on the documents set as company
root.

When moving a folder in Company to pin it, we do not want to add
the member on all children (default behavior).

Tests are adapted.

Task-4365165

X-original-commit: c9010107580c925a243266d86efced31b10c6e3d
Part-of: odoo/enterprise#75351
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>