[FIX] account_sepa_direct_debit: check commercial partner for mandate
Steps to reproduce:
- Install contacts, accountant and l10n_be
- Switch to a Belgian company
- In Accounting settings, activate "Customer Addresses" and
"SEPA Direct Debit (SDD)"
- Create a contact:
  * Country: Belgium
  * Bank Account: [a Belgian IBAN]
- Add an invoice address to the created contact
- Go to "Accounting / Customers / Direct Debit Mandates"
- Create a mandate for the created contact
- Create an invoice with the invoice address as customer
- Confirm the invoice
- Try to pay with "SEPA Direct Debit"

Issue:
No valid SEPA mandate is found for the customer.

Cause:
The mandate search is performed on the ID of the customer.
As the customer is an invoice address, he is not associated to a
bank account and a SEPA mandate.
All these data are configured for his commercial partner.

Solution:
Search the mandates based on the ID of the commercial partner.

opw-4276655

closes odoo/enterprise#75664

X-original-commit: c4fd4e6d0a46365f13a7438c1e0d7664b94bc332
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>