[MOV] appointment, *: make demo img available in appointment sub-modules
*: website_appointment

Prior to this commit, demo data images were located in
`website_appointment`, and this module was mandatory if we wanted to use
these demo images.
After this commit, these images are moved to the main `appointment`
module to be available in it and its sub-modules.

task-4317830

Part-of: odoo/enterprise#74494
Related: odoo/odoo#190946
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>