[FIX] planning: Don't use employee working hours if no resource is set
Steps to reproduce:
-------------------

- Install `Planning` module
- Go to `Planning`
- Create a new slot in the `Open Shifts` for Sunday

Issue:
------

The default `Allocated Time` is set to `00:00` while it should be set to
on the average hours per day configured on the company's calendar, which
is `08:00` by default.

Cause:
------

If no resource is assigned to the slot by default, the employee resource
is used instead. Therefore, since we have an employee on the slot, the
slot duration is computed based on the employee's working hours, instead
of the company's calendar.

Solution:
---------

Don't fallback on the employee resource if no resource is set by default

+ Remove dead code: unnecessary ternary operator

task-4353378

Part-of: odoo/enterprise#74783
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>