[IMP] stock_barcode: packaging barcodes scan as contents on homepage
Normally, scanning the barcode of a product packaging has no effect on
the barcode module homepage. For the convenience of not having to open
the packaging in order to scan the barcode of the product it contains
just to locate it, the barcode on the packaging itself is now treated
the same way as the barcode on the product itself when scanned on the
module homepage. The behaviour with inventory taking (ie adding the
product as many times as the packaging contains) remains unchanged.

Task ID: 4284938

closes odoo/enterprise#74785

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>