[FIX] website_sale_stock_renting: fix tour at night
The tour would fail after 11pm as the smallest rental interval for the
item is 1 hours leading to default dates leaking into the next day
leading to more hours than planned being encoded.
By forcing the date to be the first day of the next month we can ensure
the date is always valid.

Runbot Error 108371

closes odoo/enterprise#74910

X-original-commit: eebc00d35970acb211e07ad75def06b216892cba
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>