[IMP] sale_subscription: make set_close more coherent
Before this commit, the end_date in set_close was always set to today.
During a renewal, it was later updated. Unfortunately, it could break
customization when the parent end_date was set to today during a shrt
amount of time. During that time, it would be earlier than the
start_date and this could lead to issues on some customizations.

This change aimes to avoid setting the end_date twice and have single
update of the end_date during the renewal process.

closes odoo/enterprise#74861

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>