[FIX] spreadsheet_edition: Adapt spreadsheet thumbnail to safari
In safari (and all webkit navigators) `HTMLCanvasElement.toDataUrl` does
not support `image/webp` as `type` parameter.
https://developer.mozilla.org/en-US/docs/Web/API/HTMLCanvasElement/toDataURL#browser_compatibility

Before this commit, spreadsheet_edition/hooks.js:useSpreadsheetThumbnail
draws an image to a canvas and convert it to base64.
But the code assumes that the conversion to webp will work.

`toDataUrl` uses `png` type if it doesn't work
https://developer.mozilla.org/en-US/docs/Web/API/HTMLCanvasElement/toDataURL#type
With this commit, the regexp that removes the encoding data will supports
`webp` **and** `png`

opw-4361731

closes odoo/enterprise#74941

X-original-commit: badcdf608c489403124bf48e82c47943231e8a8a
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>