[FIX] planning: add missing run in step for onboarding tour
Before this commit, the step is not used when the tour is manually
launched since the step is considered just to be a check to make sure
the button is there in the view instead of asking to the user to click
on it.

This commit adds `click` event on that step to make sure to ask to the
user to click on the button selected inside that step.

X-original-commit: 662b8df77b9afe31ee8b031a74ceab8264737eeb
Part-of: odoo/enterprise#75071
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>