[IMP] documents_project(,_sale): remove docs links to projects/tasks
Simplify the bridge(s): stop centralizing record attachments (chatter)
in Documents.
Keep creating a folder per project and allowing to specify tags.

Links from tasks to documents will no longer be supported as
a standard feature. Users should create task-related folders
if they wish to keep this functionality. Therefore, the
probably-not-much-used action Create a Task is also removed.

A document (or a shortcut) has to be located in the project folder
to be counted in the stats buttons.
This allows to open the documents portal for project as well.

Technical note: we've removed the `documents_folder_id` `onchange`
method (logic moved to `write`) because now that all documents
inside a folder are considered linked, the consequences of
selecting a wrong folder on the project form view would have been
more dramatic (all documents inside the wrongly selected folder
would be moved in the then correctly chosen project folder).

Task-4298444

Part-of: odoo/enterprise#75214
Related: odoo/upgrade#6884