[FIX] account: make accounting fields invisble without CoA
When having Accounting installed with a company without chart template,
we can not create a partner as bot `property_account_payable_id` and
`property_account_receivable_id` are required.

With ff54822738925d5b5b952520bbfbd4a76c876809 we add a non stored computed field to handle
the `required` attribute on partner view.
We use this field to make the `Accounting Entries` group invisble when
it is not needed.

opw-4323694

closes odoo/enterprise#75320

X-original-commit: 3beab54f3d549745bec363583981812aef88d054
Related: odoo/odoo#189959
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>