[FIX] documents,marketing_automation: fix "list" prop in KanbanRecord
The documents and marketing_automation apps previously used some information
in the "list" prop that was removed in favor of a optional field "groupByField"
for simplification purposes.
See changes introduced by odoo/odoo#189109 and odoo/enterprise#74863

Now we only use the "selection" prop of the list for documents and
the code to display the confirmation message when deleting parent activities
is moved in the HierarchyKanbanRenderer with the help of the deleteRecord
function.

task-4377706

closes odoo/enterprise#75048

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>