[IMP] account_reports: use readonly server calls when possible
api.readonly allows running the queries of an rpc call in readonly mode, on a replicate of the database, in order to better spread the load. So far, accounting reports didn't use them because, in some cases, their computation requires the creation of a temporary table. We improve that by introducing a new option key telling whether the report can run in readonly mode or not, depending on the fact such temporary table is in use or not. The value of this option key then influences which end point is called to compute the report between the regular one, or an api.readonly version of it.

closes odoo/enterprise#75699

X-original-commit: 449a133a573b2d3f29dd025584536fb3accb31c2
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>