[REF] point_of_sale, pos_*: adapt generic components to PoS context.
*: l10n_cl_edi_pos, l10n_de_pos_cert, pos_avatax, pos_blackbox_be,
pos_l10n_se, pos_urban_piper

Components in the Point of Sale (PoS) context should always be able to
receive related model objects as props.

Usually, if a component is only used in PoS or Self Order, it can use
services to get the needed object. However, if a component is used in
both Self Order and PoS, it cannot use services to get the object
because the service will be different in each context.

Since Self Order and PoS both use the same models, we can pass the model
object as props to the component.

PoS context components should not be used in the customer display. Since
the customer display is a separate context, it receives pre-processed
data from PoS. Therefore, we need to create a new component for the
customer display.

closes odoo/enterprise#75055

Related: odoo/odoo#189446
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>