[IMP] web_grid: make sure `span` and `step` of ranges are valid in grid
Before this commit, `range` tag can be defined as child of field, that
range is used if the parent field is a column and a Date/Datetime field
to know the number of grid columns to display and how to fetch data to
fill in those columns. However, the content of `span` and `step`
attributes defined in the `range` xml tag is not really checked during
the xml validation of the grid view.

This commit makes sure `span` attribute only contains `day`, `week`,
`month` or `year` and `step` attribute is either `day` or `month`.

task-4373526

closes odoo/enterprise#75059

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>