[FIX] social_demo: don't duplicate dependency data
Although `social_demo` is a strong independent module it's a bit odd
that it partially duplicates a bunch of data from `base` and
`product`, and if it gets installed (possibly by mistake) into a
database for testing, some tests don't cope well with the issue.

Remove the duplicated records and just reference the base ones. Leave
`product_product_4d` as it's either novel to `social_demo` or was
removed from `product` since it was initially duplicated.

closes odoo/enterprise#75065

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>