[IMP] sale_commission: commission by quarter
Before this commit, when we groupped payment_date by quarter, the
aggreation was not correct. This commit allow to have monthly periods
and compute the payment by quarters.

Moreover, before this commit, forecast were not displayed for each
salesmen for future periods. It was not possible to easily display them.
The forecast compulation take into account each plan and salesperson to
easily update the salesperson forecast value even when they have no
achievement yet.

closes odoo/enterprise#75466

Taskid: 4376927
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>