[REF] web_grid: merge GridDataPoint into GridModel
Before this commit, the GridModel class contained a GridDataPoint to let
that additional to fetch the grid data and build the different elements
inside the grid view (rows, columns, sections, cells). This concept is a
bit complex since the developer who wants to do customisation does not
really know if he should do the customisation in GridDataPoint or
GridModel.

This commit merges the GridDataPoint into GridModel to only have one
"real model" in the grid to facilitate the eventual customisation to do
for the grid view (like the ones made for grid view of timesheets).

task-4354264

closes odoo/enterprise#75116

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>