[FIX] knowledge: allow create in external embedded view
This commit fixes an issue with the external embedded views in Knowledge where
the `New` button would raise a traceback.
The issue was introduced in the big Knowledge refactor via #67083.
This was caused by the action service when the doAction was called inside the
createRecord of the ReadonlyEmbeddedViewComponent.

The views returned by the find method called upon the action's views array were
always undefined. This meant that we never found the correct view to use which
raised the traceback. The type was also missing from the parameters.

To fix this, the arrow function given to the find method has been updated so
that it returns a correct value.
And the action is now a type `ir.actions.act_window`.

task-4367683

closes odoo/enterprise#75364

X-original-commit: 9334e7feb7f801e48648f24b261e4748b0010afc
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>