[FIX] documents: fix error when creating vendor bill from split tool
How to reproduce:
- Install documents_account
- In Document App, folder “Finance”, select 2 pdf
- Click on Action -> Merge PDFs
- Click on “Create Vendor Bill”

You get an error “psycopg2.errors.SerializationFailure: could not serialize
access due to concurrent update”

The error was caused by the update of the pdf preview during the execution of
the action “Create Vendor Bill” (which both update the document). We solve the
problem by waiting the execution of the action before executing the rest
(especially the update of the preview).

Task-4381458

X-original-commit: 23be4830b071be803801c52f32d72f8d5ba6f08f
Part-of: odoo/enterprise#75248
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Pierre-Yves Dufays (pydu) <pydu@odoo.com>