[IMP] web_studio: allow editing action without id
On a form view, navigate in a many2one.
Before this commit, the studio systray was disabled because the action had no id or xml_id.
That is to say the action had no record in database and was built synthetically.

After this commit, we allow to enter studio with an action that has no such backing record in database.

Of course, the actionEditor is mostly useless with such action.

task-4356704

closes odoo/enterprise#74800

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>