[IMP] web_gantt: copy dates explicitely
By pressing on "ctrl" key, and than drag&dropping a record in gantt
view, it copies it. If you drag&drop it in the same column, it doesn't
add its start and end dates in the default values of the `copy`.

The reason for that is to be found in `dragPillDrop` which is also
when "ctrl" is not pressed, and so the copy mode not on
(it then calls `write`). If the dates haven't changed, there
is no reason to rewrite the same values, so it makes sense to pop them.

Yet, for the "copy" flow, it is usefull to explicitely add the date
fields in the default values, even if they are the same, because the
date fields could be `copy=False`. That is what is done in this commit.

task-4295758

closes odoo/enterprise#73090

Related: odoo/odoo#185894
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>